fix(client+server): remove user controller + move export user controller to retail location controller + update client endpoint address
