+++ packages/server/src/modules/retail-location/retail-location.controller.ts
@@ -1,9 +1,11 @@
-import { createReadStream } from "node:fs";
+import { createReadStream, writeFileSync } from "node:fs";
 import { rm, writeFile } from "node:fs/promises";
 import { resolve } from "node:path";
 import {
   Controller,
   Get,
+  Header,
+  Inject,
   NotFoundException,
   Param,
   Put,
@@ -14,22 +16,27 @@ import {
 } from "@nestjs/common";
 import { FileInterceptor } from "@nestjs/platform-express";
 import { Prisma, Role, User } from "@prisma/client";
+import { upperCase } from "lodash";
+import { rootConfiguration, type RootConfiguration } from "src/config/root";
 import { AuthService } from "src/modules/auth/auth.service";
 import { CurrentUser } from "src/modules/auth/decorators/current-user.decorator";
 import { Public } from "src/modules/auth/decorators/public-route.decorator";
 import { PrismaService } from "src/modules/prisma/prisma.service";
+import { getPrismaRetailLocationFilters } from "src/modules/retail-location/retail-location.helpers";
 import { RetailLocationService } from "src/modules/retail-location/retail-location.service";
 import { Theme } from "src/modules/retail-location/theme.args";
 
-@Controller()
+@Controller("location")
 export class RetailLocationController {
   constructor(
     private readonly prisma: PrismaService,
     private readonly authService: AuthService,
     private readonly locationService: RetailLocationService,
+    @Inject(rootConfiguration.KEY)
+    private readonly rootConfig: RootConfiguration,
   ) {}
 
-  @Get("/location/:id/logo/:name")
+  @Get(":id/logo/:name")
   @Public()
   async getRetailLocationLogo(
     @Param("id") locationId: string,
@@ -63,7 +70,7 @@ export class RetailLocationController {
     );
   }
 
-  @Put("/location/:id/logo")
+  @Put(":id/logo")
   @UseInterceptors(FileInterceptor("file"))
   async uploadRetailLocationLogo(
     @Param("id") retailLocationId: string,
@@ -130,5 +137,62 @@ export class RetailLocationController {
 
   private getLogoDirectory(locationId: string) {
     return this.locationService.resolveStoragePath(locationId, "./logo");
+  }
+
+  @Get(":id/export-active-users")
+  @Header("Content-Type", "text/csv")
+  async exportActiveUsers(
+    @Param("id") retailLocationId: string,
+    @CurrentUser() currentUser: User,
+  ) {
+    await this.authService.assertMembership({
+      userId: currentUser.id,
+      message: "Forbidden access",
+      role: Role.ADMIN,
+    });
+
+    const { activeUsersFilter } =
+      getPrismaRetailLocationFilters(retailLocationId);
+
+    const users = await this.prisma.user.findMany({
+      where: { emailVerified: true, ...activeUsersFilter },
+      select: {
+        firstname: true,
+        lastname: true,
+        email: true,
+        phoneNumber: true,
+      },
+    });
+
+    if (users.length === 0) {
+      throw new UnprocessableEntityException(
+        "No active users found for the specified retail location.",
+      );
+    }
+
+    const csvHeader = Object.keys(users[0])
+      .map((value) => upperCase(value))
+      .join(",");
+    const csvBody = users
+      .map((user) =>
+        Object.values(user)
+          // Wrap values in double quotes to handle special characters
+          .map((value) => `"${value}"`)
+          .join(","),
+      )
+      .join("\n");
+
+    const filePath = this.#resolveContactsCsvFilePath(retailLocationId);
+
+    writeFileSync(filePath, `${csvHeader}\n${csvBody}`);
+
+    return new StreamableFile(createReadStream(filePath));
+  }
+
+  #resolveContactsCsvFilePath(locationId: string) {
+    return resolve(
+      this.rootConfig.storagePath,
+      `./location/${locationId}/user_list_${locationId}.csv`,
+    );
   }
 }
